examples/09_marketing_strategy/requirements.txt
crewai
crewai-tools
mcp
pyyaml
python-dotenv
langchain-core
psutil
requests
